+++ apps/v1/src/admin/pages/AdminLayout.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@plug/ui';
 import {Outlet, Link} from 'react-router-dom';
-import {useProfileStore} from "@plug/v1/auth/controller/useProfileStore";
-import {logOut} from "@plug/v1/auth/api/auth";
+import {useProfileStore} from "@plug/v1/common/auth/controller/useProfileStore";
+import {logOut} from "@plug/v1/common/auth/api/auth";
 
 const AdminLayout = () => {
     const user = useProfileStore((state) => state.user);
