+++ apps/v1/src/app/modules/view/layouts/Header/UserProfile.tsx
@@ -1,4 +1,4 @@
-import {logOut} from "@plug/v1/auth/api/auth";
+import {logOut} from "@plug/v1/common/auth/api/auth";
 
 const UserProfile = () => {
     return(
